requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
docutils==0.14
grpcio==1.21.1
grpcio-tools==1.21.1
idna==2.8
mnemonic==0.18
pbkdf2==1.3
pkginfo==1.5.0.1
protobuf==3.8.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pysha3==1.0.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
toml==0.10.0
tqdm==4.32.2
twine==1.13.0
urllib3==1.25.3
webencodings==0.5.1
